LastState, PolicyHandler, MediaManager singletons removed

+++ src/components/media_manager/include/media_manager/socket_streamer_adapter.h
@@ -42,7 +42,7 @@ namespace media_manager {
 class SocketStreamerAdapter : public StreamerAdapter {
  public:
   SocketStreamerAdapter(const std::string& ip,
-                        int32_t port,
+                        uint16_t port,
                         const std::string& header);
   virtual ~SocketStreamerAdapter();
 
@@ -51,7 +51,7 @@ class SocketStreamerAdapter : public StreamerAdapter {
    public:
     SocketStreamer(SocketStreamerAdapter* const adapter,
                    const std::string& ip,
-                   int32_t port,
+                   uint16_t port,
                    const std::string& header);
     virtual ~SocketStreamer();
 
@@ -62,7 +62,7 @@ class SocketStreamerAdapter : public StreamerAdapter {
 
    private:
     std::string ip_;
-    int32_t     port_;
+    uint16_t    port_;
     std::string header_;
 
     int32_t     socket_fd_;
